change models, change params, rename models file
